LayoutTests/platform/ios-sim-deprecated/fast/dom/custom/lifecycle-ready-paste-expected.txt

CONSOLE MESSAGE: line 9: TypeError: undefined is not a function (evaluating 'document.register')
This test ensures that the lifecycle callbacks of editing-originated elements are visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL window.callbacksCalled should be X-FOO,DIV. Was .
FAIL window.callbacksCalled should be X-FOO,DIV,X-FOO,DIV. Was .
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
